\docType{data}
\name{calflung}
\alias{calflung}
\title{calflung dataset}
\format{a data frame with 50 observations of the following 2 variables, no NAs
\describe{
\item{group}{Treatment group. One of con = control or vac = vaccinate }
\item{lesion}{Percent lung lesion, in decimal form}
}}
\description{
something here
}
\keyword{datasets}

