% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMF.R
\name{ggmf}
\alias{ggmf}
\title{ggplot2 extension for a mf object}
\usage{
ggmf(output, scale = "fixed", fit.lm = TRUE)
}
\arguments{
\item{output}{the output from \code{mf}}

\item{scale}{Are scales shared across all facets (the default, \code{"fixed"}), or do they vary across rows (\code{"free_x"}), columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{fit.lm}{If \code{TRUE}, the default, fitted line of linear model and estimated of slope are displayed.}
}
\value{
a figure for multi-functionality with uncorrelated/correlated functions.
}
\description{
\code{ggmf}: the \code{\link[ggplot2]{ggplot}} extension for \code{mf} object to plot the correlation between species richness and multi-functionality
}
\examples{
## Display fitted line of linear model
data(output1)
ggmf(output1, scale = 'fixed', fit.lm = TRUE)

## Not display fitted line of linear model
ggmf(output1, scale = 'fixed', fit.lm = FALSE)

}
