% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMF.R
\name{ggmfbeta}
\alias{ggmfbeta}
\title{ggplot2 extension for a mfbeta object}
\usage{
ggmfbeta(output, scale = "fixed", fit.lm = TRUE)
}
\arguments{
\item{output}{the output from \code{mfbeta}}

\item{scale}{Are scales shared across all facets (the default, \code{"fixed"}), or do they vary across rows (\code{"free_x"}), columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{fit.lm}{If \code{TRUE}, the default, fitted line of linear model and estimated of slope are displayed.}
}
\value{
return a figure when input contains only uncorrelated or correlated MF decomposition. returns a list of two figures when input contains both uncorrelated and correlated MF decomposition.
}
\description{
\code{ggmfbeta}: the \code{\link[ggplot2]{ggplot}} extension for \code{mfbeta} object to plot the relationship between decomposition of species richness and multi-functionality
}
\examples{
## Display fitted line of linear model
data(output2)
ggmfbeta(output2, scale = 'fixed', fit.lm = TRUE)

## Not display fitted line of linear model
ggmfbeta(output2, scale = 'fixed', fit.lm = FALSE)

}
