% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLBoot.r
\name{HLBoot}
\alias{HLBoot}
\title{Bootstrap CI for MF, HL, and Qdif}
\usage{
HLBoot(
  formula,
  data,
  vac_grp = "vac",
  con_grp = "con",
  b = 100,
  B = 100,
  alpha = 0.05,
  hpd = TRUE,
  bca = FALSE,
  return.boot = FALSE,
  trace.it = FALSE,
  seed = sample(1:1e+05, 1),
  compare = deprecated()
)
}
\arguments{
\item{formula}{Formula of the form \code{y ~ x + cluster(w)}, where y is a
continuous response, x is a factor with two levels of treatment, and w is a
factor indicating the clusters.}

\item{data}{Data frame}

\item{vac_grp}{The name of the vaccinated group.}

\item{con_grp}{The name of the control group.}

\item{b}{Number of bootstrap samples to take with each cycle}

\item{B}{Number of cycles, giving the total number of samples = B * b}

\item{alpha}{Complement of the confidence level}

\item{hpd}{Boolean whether to estimate highest density intervals for MF and
HL.}

\item{bca}{Boolean whether to estimate BCa intervals for MF.}

\item{return.boot}{Boolean whether to save the bootstrap samples of the
statistics.}

\item{trace.it}{Boolean whether to display verbose tracking of the cycles.}

\item{seed}{to initialize random number generator for reproducibility. Passed
to \code{set.seed}.}

\item{compare}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Text vector stating the factor levels: \code{compare[1]} is the
control or reference group to which \code{compare[2]} (vaccinate) is compared}
}
\value{
a \linkS4class{mfhlboot} data object
}
\description{
Estimates bootstrap confidence intervals for MF, HL, and Qdif.
}
\details{
Estimates bootstrap confidence intervals for the mitigated fraction
(MF), Hodge-Lehmann estimator (HL), and the difference of medians and
quartiles (Qdif). Equal tailed intervals are provided for all three,
highest density intervals are optionally provided for MF and HL, and BCa
intervals are optionally provided for MF.  The Hodges-Lehmann estimator is
the median difference; it assumes that the two distributions have the same
shape and differ by a constant shift. Assumes data is single pool (no
nesting).
}
\examples{
HLBoot(lesion ~ group, calflung, seed = 12345)

}
\references{
Hodges JL, Lehmann EL, (1963). Estimates of location based on
rank tests. \emph{Annals of Mathematical Statistics.} \strong{34:598--611}.

Siev D, (2005). An estimator of intervention effect on disease severity.
\emph{Journal of Modern Applied Statistical Methods.}
\strong{4:500--508}.

Efron B, Tibshirani RJ. \emph{An Introduction to the Bootstrap.} Chapman and
Hall, New York, 1993.
}
\seealso{
\linkS4class{mfhlboot}
}
\author{
\link{MF-package}
}
