% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.R
\docType{package}
\name{MF-package}
\alias{MF}
\alias{MF-package}
\title{MF: Mitigated Fraction}
\description{
Calculate MF (mitigated fraction) with clustering and bootstrap options. See http://goo.gl/pcXYVr for definition of MF. No endorsement, claim, or warranty is implied for this package. It is made available for investigational or pedagogical use only.
}
\author{
\strong{Maintainer}: CVB Statistics \email{CVB.Data.Help@usda.gov} (starting 2014)

Authors:
\itemize{
  \item David Siev (1990-2019) [former owner]
  \item Marie Vendettuoli (starting 2010)
}

Other contributors:
\itemize{
  \item Brian Woerth (starting 2018) [contributor]
}

}
\keyword{internal}
