% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{MF_measures}
\alias{MF_measures}
\title{MF_measures}
\usage{
MF_measures(
  data,
  q = c(0, 1, 2),
  fun_cols = 1:ncol(data),
  ecosystem = "single",
  species_col = NULL,
  is_normalized = TRUE,
  negative = NULL,
  is_zero_miss = TRUE
)
}
\arguments{
\item{data}{(a) For \code{is_normalized = \code{TRUE}}, data can be input as a vector of functions (for a single assemblage), matrix/data.frame (assemblages by functions). In this case, the data input must be normalized between 0 and 1 already. \cr
(b) For \code{is_normalized = \code{FALSE}}, data can only be input as a matrix/data.frame (multiple assemblages by functions).}

\item{q}{a numerical vector specifying the diversity orders. Default is 0, 1 and 2.}

\item{fun_cols}{the order number of the columns which be used as the function variable.}

\item{ecosystem}{\code{'single'} or \code{'multiple'}. MF measures in a single ecosystem or multiple ecosystems.}

\item{species_col}{(required only when \code{ecosystem = 'multiple'}), name of the column to be used to calculate the species diversity.}

\item{is_normalized}{a logical variable to represent whether the data input is normalized between 0 and 1, default is \code{TRUE}.}

\item{negative}{(required only when \code{is_normalized = FALSE}), name of columns to be normalized negatively.}

\item{is_zero_miss}{(required only when \code{is_normalized = FALSE}), a logical variable to determine whether the zero in data represents no value or not, default is \code{TRUE}..}
}
\value{
a data.frame
}
\description{
\code{MF_measures}:\cr
Multi-functionality measures in a single ecosystem or multiple ecosystems. This is a function in the framework of Hill-Chao numbers
}
