% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFBoot.r
\name{MFBoot}
\alias{MFBoot}
\title{Bootstrap MF CI}
\usage{
MFBoot(
  formula,
  data,
  vac_grp = "vac",
  con_grp = "con",
  b = 100,
  B = 100,
  alpha = 0.05,
  hpd = TRUE,
  bca = FALSE,
  return.boot = FALSE,
  trace.it = FALSE,
  seed = sample(1:1e+05, 1),
  compare = deprecated()
)
}
\arguments{
\item{formula}{Formula of the form \code{y ~ x}, where y is a continuous response
and x is a factor with two levels.}

\item{data}{Data frame}

\item{vac_grp}{The name of the vaccinated group.}

\item{con_grp}{The name of the control group.}

\item{b}{Number of bootstrap samples to take with each cycle}

\item{B}{Number of cycles, giving the total number of samples = B * b}

\item{alpha}{Complement of the confidence level}

\item{hpd}{Estimate highest density intervals?}

\item{bca}{Estimate BCa intervals?}

\item{return.boot}{Save the bootstrap sample of the MF statistic?}

\item{trace.it}{Verbose tracking of the cycles?}

\item{seed}{to initialize random number generator for reproducibility. Passed
to \code{set.seed}.}

\item{compare}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Text vector stating the factor levels:
\code{compare[1]} is the control or reference group to which \code{compare[2]}
(vaccinate) is compared}
}
\value{
a \linkS4class{mfboot} data object
}
\description{
Estimates bootstrap confidence intervals for the mitigated
fraction.
}
\details{
Resamples the data and produces bootstrap confidence intervals.
Equal tailed intervals are estimated by the percentile method. Highest
density intervals are estimated by selecting the shortest of all possible
intervals. For BCa intervals, see Efron and Tibshirani section 14.3.
}
\examples{

MFBoot(lesion ~ group, calflung, seed = 12345)

}
\references{
Siev D. (2005). An estimator of intervention effect on disease
severity. \emph{Journal of Modern Applied Statistical Methods.}
\strong{4:500--508}

Efron B, Tibshirani RJ. \emph{An Introduction to the Bootstrap.} Chapman and
Hall, New York, 1993.
}
\seealso{
\linkS4class{mfboot}
}
\author{
\link{MF-package}
}
