% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFClus.r
\name{MFClus}
\alias{MFClus}
\title{Clustered mitigated fraction}
\usage{
MFClus(
  formula,
  data,
  vac_grp = "vac",
  con_grp = "con",
  trace.it = FALSE,
  compare = deprecated()
)
}
\arguments{
\item{formula}{Formula of the form \code{y ~ x + cluster(w)}, where y is a
continuous response, x is a factor with two levels of treatment, and w is a
factor indicating the clusters.}

\item{data}{Data frame.  See \code{Note} for handling of input data with more than
two levels.}

\item{vac_grp}{The name of the vaccinated group.}

\item{con_grp}{The name of the control group.}

\item{trace.it}{Verbose tracking of the cycles? Default FALSE.}

\item{compare}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Text vector stating the factor levels: \code{compare[1]} is the
control or reference group to which \code{compare[2]} (vaccinate) is compared}
}
\value{
a \linkS4class{mfcluster} data object
}
\description{
Estimates mitigated fraction from clustered or stratified data.
}
\details{
Averages the U statistic over the clusters and computes MF from it. Clusters
are excluded if they do not include both treatments.
}
\note{
If input data contains more than two levels of treatment, rows
associated with unused treatment levels will be removed.

Factor levels for treatments not present in the input data will be ignored.

Clusters with missing treatments will be excluded. See
\linkS4class{mfbootcluster} or use \code{trace.it} to identify excluded clusters.
}
\examples{
\dontrun{
MFClus(lesion ~ group + cluster(litter), piglung)
}
}
\references{
Siev D. (2005). An estimator of intervention effect on disease
severity. \emph{Journal of Modern Applied Statistical Methods.}
\strong{4:500--508}
}
\seealso{
\linkS4class{mfcluster}
}
\author{
\link{MF-package}
}
