% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFClusBoot.r
\name{MFClusBoot}
\alias{MFClusBoot}
\title{Bootstrap MF CI from clustered data}
\usage{
MFClusBoot(
  formula,
  data,
  vac_grp = "vac",
  con_grp = "con",
  boot.cluster = TRUE,
  boot.unit = TRUE,
  b = 100,
  B = 100,
  alpha = 0.05,
  hpd = TRUE,
  return.boot = FALSE,
  trace.it = FALSE,
  seed = sample(1:1e+05, 1),
  compare = deprecated()
)
}
\arguments{
\item{formula}{Formula of the form \code{y ~ x + cluster(w)}, where y is a
continuous response, x is a factor with two levels of treatment, and w is a
factor indicating the clusters.}

\item{data}{Data frame. See \code{Note} for handling of input data with more than
two levels.}

\item{vac_grp}{The name of the vaccinated group.}

\item{con_grp}{The name of the control group.}

\item{boot.cluster}{Boolean whether to resample the clusters.}

\item{boot.unit}{Boolean whether to resample the units within cluster.}

\item{b}{Number of bootstrap samples to take with each cycle}

\item{B}{Number of cycles, giving the total number of samples = B * b}

\item{alpha}{Complement of the confidence level}

\item{hpd}{Boolean whether to estimate highest density intervals.}

\item{return.boot}{Boolean whether to save the bootstrap sample of the MF
statistic.}

\item{trace.it}{Boolean whether to display verbose tracking of the cycles.}

\item{seed}{to initialize random number generator for reproducibility. Passed
to \code{set.seed}.}

\item{compare}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Text vector stating the factor levels:
\code{compare[1]} is the control or reference group to which \code{compare[2]}
(vaccinate) is compared}
}
\value{
a \linkS4class{mfbootcluster} data object
}
\description{
Estimates bootstrap confidence intervals for the mitigated
fraction from clustered or stratified data.
}
\details{
Resamples the data and produces bootstrap confidence intervals.
Equal tailed intervals are estimated by the percentile method. Highest
density intervals are estimated by selecting the shortest of all possible
intervals.
}
\note{
If input data contains more than two levels of treatment, rows
associated with unused treatment levels will be removed.

Factor levels for treatments not present in the input data will be ignored.

Clusters with missing treatments will be excluded. See
\linkS4class{mfbootcluster} or use \code{trace.it} to identify excluded clusters.
}
\examples{
MFClusBoot(lesion ~ group + cluster(litter), piglung, seed = 12345)
}
\references{
Siev D. (2005). An estimator of intervention effect on disease
severity. \emph{Journal of Modern Applied Statistical Methods.}
\strong{4:500--508}

Efron B, Tibshirani RJ. \emph{An Introduction to the Bootstrap.} Chapman and
Hall, New York, 1993.
}
\author{
\link{MF-package}
}
