% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{mfclusboothier-class}
\alias{mfclusboothier-class}
\alias{mfclusboothier}
\title{Class mfclusboothier}
\usage{
mfclusboothier$new(MFhBoot, MFnestBoot)
}
\description{
Class mfclusboothier is created from output of function 
MFClusBootHier.
}

\author{
Marie Vendettuoli \email{marie.c.vendettuoli@aphis.usda.gov}
}
\keyword{documentation}
