% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{mfclushier-class}
\alias{mfclushier-class}
\alias{mfclushier}
\title{Class mfclushier}
\usage{
mfclushier$new(MFh, MFnest)
}
\description{
Class mfclushier is created from output of function MFClusHier
}

\section{Fields}{

\itemize{
\item{\code{MFh: }}{output from MFh. A \code{\link{mfhierdata}} object.}
\item{\code{MFnest: }}{output from MFnest. A tibble.}
}
}

\seealso{
\code{\link{MFh}}, \code{\link{MFnest}}
}
\author{
Marie Vendettuoli \email{marie.c.vendettuoli@aphis.usda.gov}
}
\keyword{documentation}
