% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFHier-wrappers.r
\name{MFClusHier}
\alias{MFClusHier}
\title{MFClusHier}
\usage{
MFClusHier(formula, data, compare = c("con", "vac"), which.factor = "All")
}
\arguments{
\item{formula}{Formula of the form y ~ x + a/b/c, where y is a continuous
response, x is a factor with two levels of treatment, and a/b/c are
grouping variables corresponding to the clusters. Nesting is assumed to be
in order, left to right, highest to lowest. So a single level of "a" will
contain multiple levels of "b" and a single level of "b" will contain
multiple levels of "c".}

\item{data}{a data.frame or tibble with the variables specified in formula.
Additional variables will be ignored.}

\item{compare}{Text vector stating the factor levels - \code{compare[1]} is the
control or reference group to which \code{compare[2]} is compared.}

\item{which.factor}{one or more variable(s) of interest. This can be any of
the core or nest variables from the data set. If none or NULL is specified,
MF will be calculated for the whole tree.}
}
\value{
A list with the following elements:
\itemize{
\item \code{MFh}: as output from \link{MFh}.
\item \code{MFnest}: as output from \link{MFnest}.
}
}
\description{
Calculate mitigated fraction directly from hierarchial nested
data. Combines \link{MFh} and \link{MFnest} into a single function.
}
\note{
\code{Core} variable is the variable corresponding to the lowest nodes of
the hierarchical tree. \code{Nest} variables are those above the core. \code{All}
refers to a summary of the entire tree.
}
\examples{
a <- data.frame(
room = paste("Room", rep(c("W", "Z"), each = 24)),
pen = paste("Pen", rep(LETTERS[1:6], each = 8)),
litter = paste("Litter", rep(11:22, each = 4)),
tx = rep(rep(c("vac", "con"), each = 2), 12))
set.seed(76153)
a$lung[a$tx == "vac"] <- rnorm(24, 5, 1.3)
a$lung[a$tx == "con"] <- rnorm(24, 7, 1.3)
thismf <- MFClusHier(lung ~ tx + room / pen / litter, a)
thismf$MFnest
aCore <- thismf$MFh
aCore
aCore$data
aCore$formula
aCore$compare
}
\seealso{
\link{MFh}, \link{MFnest}
}
\author{
\link{MF-package}
}
