% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFClusHier.R
\name{MFClusNested}
\alias{MFClusNested}
\title{Clustered mitigated fraction for nested experimental design.}
\usage{
MFClusNested(formula, data, compare = c("con", "vac"), trace.it = FALSE)
}
\arguments{
\item{formula}{Formula of the form \code{y ~ x + cluster(w1/w2/..)}, where y is a 
continuous response, x is a factor with two levels of treatment, and w1/w2/... are  
variables indicating the nested clusters.}

\item{data}{Data frame.}
}
\author{
David Siev \email{david.siev@aphis.usda.gov}
}
\references{
Siev D. (2005). An estimator of intervention effect on disease severity.
\emph{Journal of Modern Applied Statistical Methods.} \bold{4:500--508}
}
\seealso{
\code{\link{mfcluster-class}}, \code{\link{MFClus}}
}

