% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFnestBoot.r
\name{MFhBoot}
\alias{MFhBoot}
\title{MFhBoot}
\usage{
MFhBoot(formula, data, compare = c("con", "vac"), nboot = 10000,
  alpha = 0.05, boot.unit = TRUE, boot.cluster = TRUE)
}
\description{
MFh using bootstrapping
}
\examples{
set.seed(76153)
a <- data_frame(room = paste('Room', rep(c('W','Z'), each = 24)),
                pen = paste('Pen', rep(LETTERS[1:6], each = 8)),
                litter = paste('Litter', rep(11:22, each = 4)),
                tx = rep(rep(c('vac', 'con'), each = 2), 12)) \%>\%
  mutate(lung = ifelse(tx == 'vac', rnorm(24, 5, 1.3), rnorm(24, 7, 1.3)))
a

formula <- lung ~ tx + room/pen/litter
nboot <- 10000
boot.cluster <- TRUE
boot.unit <- TRUE
which.factors <- c('All', 'room', 'pen', 'litter')
set.seed(12345)

system.time(test1 <- MFhBoot(formula, a, 
                            nboot = 10000,
                             boot.cluster = TRUE, boot.unit = TRUE))
test1$bootdat
}
