% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{mfhlboot-class}
\alias{mfhlboot-class}
\alias{mfhlboot}
\title{Class mfhlboot}
\description{
class for data objects produced by HLBoot, contains class mf
with additional fields \emph{MFstat}, \emph{HLstat}, \emph{QDIFstat}, \emph{QXstat}, \emph{QYstat}
}

\section{Fields}{

\itemize{
\item \code{nboot}: Numeric value specifying number of samples.
\item \code{alpha}: Numeric value specifying complement of confidence interval.
\item \code{seed}: Vector of integers specifying seed for pseudo-random number generator used.
\item \code{compare}: Vector of character strings naming groups compared.
\item \code{rng}: Character string naming type of random number generator.
\item \code{sample}: The bootstrapped values.
\item \code{MFstat}: Matrix with columns \emph{observed}, \emph{median}, \emph{lower}, \emph{upper} for
Equal Tailed and Highest Density estimates of mitigated fraction (MF).
\item \code{HLstat}: Matrix with columns \emph{observed}, \emph{median}, \emph{lower}, \emph{upper} for
Equal Tailed and Highest Density estimates of Hodge-Lehmann estimator (HL).
\item \code{QDIFstat}: Matrix with columns \emph{observed}, \emph{median}, \emph{lower}, \emph{upper} for
estimates of Quartile Differences.
\item \code{QXstat}: Matrix with columns \emph{observed}, \emph{median}, \emph{lower}, \emph{upper} for
quartiles of treatments, equal tailed.
\item \code{QYstat}: Matrix with columns \emph{observed}, \emph{median}, \emph{lower}, \emph{upper} for
quartiles of response, equal tailed.
}
}

\section{Contains}{
 \linkS4class{mf}
}

\seealso{
\link{HLBoot}

Other mf: 
\code{\link{mf-class}},
\code{\link{mfboot-class}},
\code{\link{mfbootcluster-class}}
}
\author{
\link{MF-package}
}
\concept{mf}
\keyword{documentation}
