% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFmp.r
\name{MFmp}
\alias{MFmp}
\title{Mitigated fraction from matched pairs}
\usage{
MFmp(
  formula = NULL,
  data = NULL,
  vac_grp = "vac",
  con_grp = "con",
  x = NULL,
  alpha = 0.05,
  df = NA_real_,
  tdist = TRUE,
  compare = deprecated()
)
}
\arguments{
\item{formula}{Formula of the form \code{y ~ x + cluster(w)}, where y is a
continuous response, x is a factor with two levels of treatment, and w is a
factor indicating the clusters.}

\item{data}{Data frame}

\item{vac_grp}{The name of the vaccinated group.}

\item{con_grp}{The name of the control group.}

\item{x}{Trinomial vector \eqn{\{\Sigma I(x < y), \Sigma I(x = y), \Sigma I(x
> y)\}}}

\item{alpha}{Complement of the confidence level.}

\item{df}{Degrees of freedom. Default N-2}

\item{tdist}{Use quantiles of t or Gaussian distribution for confidence
interval? Default t distribution.}

\item{compare}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Text vector stating the factor levels:
\code{compare[1]} is the control or reference group to which \code{compare[2]}
(vaccinate) is compared}
}
\value{
a \linkS4class{mfmp} data object
}
\description{
Estimates mitigated fraction from matched pairs.
}
\details{
Estimates \emph{MF} from matched pairs by the difference of multinomial
fractions \eqn{(\Sigma I(x<y) - \Sigma I(x > y)) / N}. The trinomial vector
is \eqn{\{\Sigma I(x<y), \Sigma I(x = y), \Sigma I(x > y)\}}
}
\note{
upper confidence interval is truncated to 1; lower confidence interval
is truncated to -1. Point estimate of 1.0 indicates complete separation.
}
\examples{
MFmp(les ~ tx + cluster(cage), mlesions)
MFmp(x = c(12, 12, 2))
}
\references{
Siev D. (2005). An estimator of intervention effect on disease
severity. \emph{Journal of Modern Applied Statistical Methods.}
\strong{4:500--508}
}
\seealso{
\linkS4class{mfmp}
}
\author{
\link{MF-package}
}
