% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFClusHier.r
\name{MFnest}
\alias{MFnest}
\title{Summations to calculate the MF for nested data from a rank table.}
\usage{
MFnest(Y, which.factor = "All")
}
\arguments{
\item{Y}{rank table (tibble or data.frame), structured as \verb{$coreTbl} output
from \link{MFh} or returned object from \url{MFh}.}

\item{which.factor}{one or more grouping variable(s) of interest. This can be
any of the core or nest variables from the data set. If none or \code{All} is
specified, a summary MF will be calculated for the whole tree.}
}
\value{
A tibble with each unique level of a variable as a row. Other values
include:
\itemize{
\item \code{MF}: Mitigated fraction for the particular level of the
variable in this row.
\item \code{N1N2}: Sum of the \code{n1n2} variable in \verb{$coreTbl} field of
\link{mfhierdata} object output by \link{MFh} for this particular variable-level
combination.
\item \code{U}: Sum of u variable in \verb{$coreTbl} field of \link{mfhierdata} object
output by \link{MFh} for this particular variable-level combination.
\item \verb{_N}: Sum of the \verb{_n} variable in \verb{$coreTbl} field of \link{mfhierdata}
object output by \link{MFh} for this particular variable-level combination.
\item \verb{_medResp}: Median of observed responses for each comparison group
for this particular variable-level combination.
}
}
\description{
Summations to calculate the MF for nested data from a rank table.
}
\note{
Core variable is the variable corresponding to the lowest nodes of the
hierarchial tree. Nest variables are those above the core. All refers to a
summary of the entire tree.
}
\examples{
a <- data.frame(
 room = paste("Room", rep(c("W", "Z"), each = 24)),
 pen = paste("Pen", rep(LETTERS[1:6], each = 8)),
 litter = paste("Litter", rep(11:22, each = 4)),
 tx = rep(rep(c("vac", "con"), each = 2), 12))
set.seed(76153)
a$lung[a$tx == "vac"] <- rnorm(24, 5, 1.3)
a$lung[a$tx == "con"] <- rnorm(24, 7, 1.3)

aCore <- MFh(lung ~ tx + room / pen / litter, a)
MFnest(aCore)

MFnest(aCore$coreTbl)

MFnest(aCore, "room")

MFnest(aCore, "pen")

MFnest(aCore, c("All", "litter"))

MFnest(aCore, "litter")

MFnest(aCore, c("room", "pen", "litter"))
}
\seealso{
\link{MFh}
}
\author{
\link{MF-package}
}
