% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFnestBoot.r
\name{MFnestBoot}
\alias{MFnestBoot}
\title{MFnestBoot}
\usage{
MFnestBoot(formula, data, compare = c("con", "vac"), B = 100, b = 100,
  alpha = 0.05, hpd = TRUE, boot.unit = TRUE, boot.cluster = TRUE)
}
\arguments{
\item{formula}{Formula of the form y ~ x + a/b/c, where y is a continuous 
response, x is a factor with two levels of treatment, and a/b/c are variables 
corresponding to the clusters. It is expected that levels of "c" are nested within 
levels of "b". Nesting is assumed to be in order, left to right, highest to lowest.}

\item{data}{a data.frame or tibble with the variables specified in formula. 
Additional variables will be ignored.}

\item{compare}{Text vector stating the factor levels - \code{compare[1]} is the control 
or reference group to which \code{compare[2]} is compared.}

\item{B}{Number of cycles, giving the total number of samples = B * b}

\item{b}{Number of bootstrap samples to take with each cycle}

\item{alpha}{complement of the confidence level}

\item{hpd}{boolean whether to estimate highest density intervals}

\item{boot.unit}{Boolean whether to resample the units within each cluster}

\item{boot.cluster}{Boolean whether to resample the clusters}

\item{compare}{Text vector stating the factor levels - compare[1] is the control or 
reference group to which compare[2] is compared.}
}
\description{
Bootstrapping MF estimate for nested heirarchy
}
\note{
B * b notation is used for consistency with MFClus
}
\examples{
a <- as_tibble(
 room = paste('Room',rep(c('W','Z'),each=24)),
 pen = paste('Pen',rep(LETTERS[1:6],each=8)),
 litter = paste('Litter',rep(11:22,each=4)),
 tx = rep(rep(c('vac','con'),each=2),12)
 )
set.seed(76153)
a$lung[a$tx=='vac'] <- rnorm(24,5,1.3)
a$lung[a$tx=='con'] <- rnorm(24,7,1.3)
MFnestBoot(lung ~ tx + room/pen/litter, a)
MFnestBoot(lung ~ tx + room/pen/litter, a, boot.unit = FALSE)
MFnestBoot(lung ~ tx + room/pen/litter, a, which.factor = 'litter')
MFnestBoot(lung ~ tx + room/pen/litter, a, n.boot = 2)
MFnestBoot(lung ~ tx + room/pen/litter, a, n.boot = 144, boot.cluster = FALSE)

}
\seealso{
\code{\link{MFClus}}
}
\author{
CVB Statistics \email{CVB.Data.Help@aphis.usda.gov}
}
