% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSubj.r
\name{MFSubj}
\alias{MFSubj}
\title{Subject components of mitigated fraction}
\usage{
MFSubj(formula, data, vac_grp = "vac", con_grp = "con", compare = deprecated())
}
\arguments{
\item{formula}{Formula of the form \code{y ~ x}, where y is a continuous response
and x is a factor with two levels}

\item{data}{Data frame}

\item{vac_grp}{The name of the vaccinated group.}

\item{con_grp}{The name of the control group.}

\item{compare}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Text vector stating the factor levels:
\code{compare[1]} is the control or reference group to which \code{compare[2]}
(vaccinate) is compared}
}
\value{
a \linkS4class{mfcomponents} data object
}
\description{
Estimates the subject components of the mitigated fraction.
}
\details{
The mitigated fraction is an estimator that quantifies an intervention's
effect on reducing the severity of a condition. Since its units are on the
probability scale, it is often a good idea to accompany it with an estimator
on the original scale of measurement.

The subject components are the individual contributions of the treated
subjects to \emph{MF}, which is the average of the subject components.
}
\examples{
x <- MFSubj(lesion ~ group, calflung)
x

mean(x$subj[, "mf_j"])

}
\references{
Siev D. (2005). An estimator of intervention effect on disease
severity. \emph{Journal of Modern Applied Statistical Methods.}
\strong{4:500--508}
}
\author{
\link{MF-package}
}
