\docType{data}
\name{piglung}
\alias{piglung}
\title{piglung dataset}
\format{a data frame with 102 observations of the following 3 variables, no NAs
\describe{
\item{lesion}{Percent lung lesion}
\item{group}{Treatment group. One of 'con' or 'vac'}
\item{litter}{Litter ID}
}}
\description{
something here
}
\keyword{datasets}

