% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF-package.r
\docType{data}
\name{calflung}
\alias{calflung}
\alias{calflung-data}
\title{calflung dataset}
\format{
a data frame with 50 observations of the following 2 variables, no
NAs
\itemize{
\item \code{group}: Treatment group. One of con = control or vac = vaccinate
\item \code{lesion}: Fraction of lungs with gross lesions.
}
}
\description{
Post-mortem examination of the lungs of groups of calves.
}
\keyword{datasets}
