% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{mf-class}
\alias{mf-class}
\alias{mf}
\title{Class mf}
\usage{
mf$new(nboot, alpha, seed, compare, rng)
}
\description{
Parent class for package MF data objects.
}

\section{Fields}{
 
\itemize{
\item{\code{nboot: } }{numeric value specifying number of samples}
\item{\code{alpha: }}{numeric value specifying complement of confidence interval}
\item{\code{seed: }}{vector of integers specifying seed for pseudo-random number generator used}
\item{\code{compare: }}{vector of character strings naming groups compared}
\item{\code{rng: }}{character string naming type of random number generator}
}
}

\seealso{
Other mf: \code{\link{mfboot-class}},
  \code{\link{mfbootcluster-class}},
  \code{\link{mfhlboot-class}}
}
\author{
Marie Vendettuoli \email{marie.c.vendettuoli@aphis.usda.gov}
}
\keyword{documentation}
