% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF-package.r
\docType{package}
\name{MF-package}
\alias{MF-package}
\alias{MF}
\title{MF Package}
\description{
Includes functions related to mitigated fraction. \cr \cr For internal use only at
the USDA Center for Veterinary Biologics. \cr
}
\details{
\tabular{ll}{
Package: \tab MF-package\cr
Type: \tab Package\cr
Version: \tab 4.3.5\cr
Date: \tab 21-Dec-2018\cr
License: \tab MIT \cr
LazyLoad: \tab yes\cr
}
}
\examples{
#---------------------------------------------
# Checking MF package
#---------------------------------------------
example(MFr)
#---------------------------------------------
# End examples
#---------------------------------------------
invisible()
}
\author{
David Siev \email{David.Siev@aphis.usda.gov}
}
