% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF.r
\docType{package}
\name{MF-package}
\alias{MF}
\alias{MF-package}
\title{MF: Mitigated Fraction}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate MF (mitigated fraction) with clustering and bootstrap options. See http://goo.gl/pcXYVr for definition of MF. No endorsement, claim, or warranty is implied for this package. It is made available for investigational or pedagogical use only.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ABS-dev/MF/}
  \item \url{https://abs-dev.github.io/MF/}
  \item Report bugs at \url{https://github.com/ABS-dev/MF/issues/}
}

}
\author{
\strong{Maintainer}: CVB Statistics \email{CVB.Data.Help@usda.gov} [former owner]

Authors:
\itemize{
  \item David Siev (until 2019)
  \item Marie Vendettuoli (until 2019)
  \item Thomas Kent (starting 2020)
}

}
\keyword{internal}
