\docType{package}
\name{MF-package}
\alias{MF}
\alias{MF-package}
\title{MF Package}
\description{
Includes functions related to mitigated fraction. \cr \cr
For internal use only at the USDA Center for Veterinary
Biologics. \cr
}
\details{
\tabular{ll}{ Package: \tab MF-package\cr Type: \tab
Package\cr Version: \tab 4.3.3\cr Date: \tab 2014-08-11\cr
License: \tab MIT \cr LazyLoad: \tab yes\cr }
}
\examples{
#---------------------------------------------
# Checking MF package
#---------------------------------------------
example(MFr)
#---------------------------------------------
# End examples
#---------------------------------------------
invisible()
}
\author{
David Siev \email{David.Siev@aphis.usda.gov}
}

