% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{mfclusboothier-class}
\alias{mfclusboothier-class}
\alias{mfclusboothier}
\title{Class mfclusboothier}
\description{
Class mfclusboothier is created from output of function
MFClusBootHier.
}

\section{Fields}{

\itemize{
\item{\code{MFhBoot: }}{output from MFhBoot. A list.}
\item{\code{MFnestBoot: }}{output from MFnestBoot. A list.}
}
}

\seealso{
\code{\link{MFhBoot}}, \code{\link{MFnestBoot}}
}
\author{
\link{MF-package}
}
\concept{mfclusboothier}
\keyword{documentation}
