% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{mfclushier-class}
\alias{mfclushier-class}
\alias{mfclushier}
\title{Class mfclushier}
\description{
Class mfclushier is created from output of function MFClusHier
}

\section{Fields}{

\itemize{
\item \code{MFh}: output from MFh. A \link{mfhierdata} object.
\item \code{MFnest}: output from MFnest. A tibble.
}
}

\seealso{
\link{MFh}, \link{MFnest}
}
\author{
\link{MF-package}
}
\concept{mfclushier}
\keyword{documentation}
