% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{mfcluster-class}
\alias{mfcluster-class}
\alias{mfcluster}
\title{Class mfcluster}
\description{
Class mfcluster is created from output of function MFClus
}

\section{Fields}{

\itemize{
\item \code{All}: vector with elements
\itemize{
\item \code{w}: Wilcoxon statistic
\item \code{u}: Mann-Whitney statistic
\item \code{r}: mean ridit
\item \code{n1}: size of group 1
\item \code{n2}: size of group 2
\item \code{mf}: mitigated fraction
}
\item \code{byCluster}: As for All, by clusters
\item \code{excludedClusters}: character vector naming clusters excluded because of missing treatment
\item \code{call}: the call to \code{MFClus}
\item \code{compare}: character vector naming groups compared
}
}

\seealso{
\link{MFClus}
}
\author{
\link{MF-package}
}
\concept{mfcluster}
\keyword{documentation}
