% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{mfcomponents-class}
\alias{mfcomponents-class}
\alias{mfcomponents}
\title{Class mfcomponents}
\description{
Class mfcomponents is created from output of function MFSubj
}

\section{Fields}{

\itemize{
\item \code{mf}: numeric estimator for mitigated fraction
\item \code{x}: numeric vector containing responses of group 1
\item \code{y}: numeric vector containing responses of group 2
\item \code{subj}: matrix where \code{mf_j}re the subject components
\item \code{compare}: character vector naming groups being compared
}
}

\seealso{
\link{MFSubj}
}
\author{
\link{MF-package}
}
\concept{mfcomponents}
\keyword{documentation}
