\docType{class}
\name{mfcomponents-class}
\alias{mfcomponents-class}
\title{Class mfcomponents}
\usage{
mfcomponents$new(mf, x, y, subj, compare)
}
\description{
Class mfcomponens is created from output of function MFSubj
}
\section{Fields}{
  \itemize{ \item{\code{mf: }}{numeric estimator for
  mitigated fraction} \item{\code{x: }}{numeric vector
  containing responses of group 1} \item{\code{y:
  }}{numeric vector containing responses of group 2}
  \item{\code{subj: }}{matrix where \code{mf.j} are the
  subject components} \item{\code{compare: }}{character
  vector naming groups being compared} }
}
\author{
Marie Vendettuoli
\email{marie.c.vendettuoli@aphis.usda.gov}
}
\seealso{
\code{\link{MFSubj}}
}
\keyword{documentation}

