% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFClusHier.r
\name{MFh}
\alias{MFh}
\title{Identify ranks for use when evaluating MF for nested hierarchy.}
\usage{
MFh(formula, data, vac_grp = "vac", con_grp = "con", compare = deprecated())
}
\arguments{
\item{formula}{Formula of the form y ~ x + a/b/c, where y is a continuous
response, x is a factor with two levels of treatment, and a/b/c are
vgrouping variables corresponding to the clusters. Nesting is assumed to be
in order, left to right, highest to lowest. So a single level of "a" will
contain multiple levels of "b" and a single level of "b" will contain
multiple levels of "c".}

\item{data}{a data.frame or tibble with the variables specified in formula.
Additional variables will be ignored.}

\item{vac_grp}{The name of the vaccinated group.}

\item{con_grp}{The name of the control group.}

\item{compare}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Text vector stating the factor levels:
\code{compare[1]} is the control or reference group to which \code{compare[2]}
(vaccinate) is compared.}
}
\value{
A \link{mfhierdata} object, which is a list of three items.
\itemize{
\item \code{coreTbl} A \link{tibble} with one row for each unique core level showing
values for:
\itemize{
\item \code{con_n} & \code{vac_n}: counts of observations for each treatment level
in the core level.
\item \code{con_medResp} & \code{vac_medResp}: median of the \code{y} continuous
response for each treatment level.
\item \code{n1n2}: product of the counts, \code{con_n} \eqn{\times} \code{vac_n}.
\item \code{w}: Wilcoxon statistic
\item \code{u}: Mann-Whitney statistic
}
\item \code{data}: A \link{tibble} of the restructured input data used for
calculations.
\item \code{compare}: The compare variables as input by user.
\item \code{formula}: The formula as input by user.
}
}
\description{
Identify ranks for use when evaluating MF for nested hierarchy.
}
\note{
Core variable is the variable corresponding to the lowest nodes of the
hierarchial tree. Nest variables are those above the core.
}
\examples{
a <- data.frame(
 room   = paste("Room", rep(c("W", "Z"), each = 24)),
 pen    = paste("Pen", rep(LETTERS[1:6], each = 8)),
 litter = paste("Litter", rep(11:22, each = 4)),
 tx     = rep(rep(c("vac", "con"), each = 2), 12))
set.seed(76153)
a$lung[a$tx == "vac"] <- rnorm(24, 5, 1.3)
a$lung[a$tx == "con"] <- rnorm(24, 7, 1.3)

aCore <- MFh(lung ~ tx + room / pen / litter, a)
aCore
}
\seealso{
\link{MFnest} for calculation of MF for nest, core and all variables.
\link{mfhierdata} for returned object. \link{MFClusHier} for a wrapper.
}
\author{
\link{MF-package}
}
