% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{mfmp-class}
\alias{mfmp-class}
\alias{mfmp}
\title{Class mfmp}
\usage{
mfmp$new(ci, x, what, alpha, tdist, df)
}
\description{
Class mfmp is created from output of function MFmp
}

\section{Fields}{

\itemize{
\item{\code{ci:} }{numeric vector of point and interval estimates}
\item{\code{x: } }{numeric vector of length three holding data}
\item{\code{what: }}{text string describing interval type}
\item{\code{alpha: }}{numeric value specifying complement of confidence interval}
\item{\code{tdist: }}{Logical indicating if t distribution(TRUE) or gaussian (FALSE)}
\item{\code{df: }}{numeric value indicating degrees freedom}
}
}

\seealso{
\code{\link{MFmp}}
}
\author{
Marie Vendettuoli \email{marie.c.vendettuoli@aphis.usda.gov}
}
\keyword{documentation}
