% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{mfmp-class}
\alias{mfmp-class}
\alias{mfmp}
\title{Class mfmp}
\description{
Class \code{mfmp} is created from output of function \code{MFmp}
}

\section{Fields}{

\itemize{
\item \code{ci}: numeric vector of point and interval estimates
\item \code{x}: numeric vector of length three holding data
\item \code{what}: text string describing interval type
\item \code{alpha}: numeric value specifying complement of confidence interval
\item \code{tdist}: Logical indicating if t distribution(TRUE) or Gaussian (FALSE)
\item \code{df}: numeric value indicating degrees freedom
}
}

\seealso{
\link{MFmp}
}
\author{
\link{MF-package}
}
\concept{mfmp}
\keyword{documentation}
