% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFClusHier.R
\name{MFnest}
\alias{MFnest}
\title{Summations to calculate the MF for nested data from a rank table.}
\usage{
MFnest(Y, which.factor = NULL)
}
\arguments{
\item{Y}{rank table, structured as \code{$coreTbl} output from \code{\link{MFh}} or
output list from MFh}

\item{which.factor}{one or more variable(s) of interest. This can be any of the core or nest variables
from the data set. If none or \code{NULL} is specified, MF will be calculated for the whole 
tree.}
}
\value{
data.frame with each unique level of a variable as a row. Values \code{N}, 
\code{U}, \code{MF} and median of responses returned.
}
\note{
Core variable is the variable corresponding to the lowest nodes of the hierarchial 
tree. Nest variables are those above the core. All refers to a summary of the entire tree.
}
\examples{
MFnest(aCore)
#   variable level  N  U    MF median_resp:con median_resp:vac
# 1      All   All 48 45 0.875        7.244466        4.909263

MFnest(aCore$coreTbl)
# Skipping median summary, no response data provided.
#   variable level  N  U    MF
# 1      All   All 48 45 0.875

MFnest(aCore, 'room')
#  variable  level  N  U        MF median_resp:con median_resp:vac
# 1     room Room W 24 22 0.8333333        7.792846        4.846574
# 2     room Room Z 24 23 0.9166667        6.708765        4.978941

MFnest(aCore, 'pen')
#   variable level N U   MF median_resp:con median_resp:vac
# 1      pen Pen A 8 6 0.50        6.792419        4.237541
# 2      pen Pen B 8 8 1.00        8.113648        5.594649
# 3      pen Pen C 8 8 1.00        7.688790        4.846574
# 4      pen Pen D 8 7 0.75        6.097081        4.978941
# 5      pen Pen E 8 8 1.00        6.858526        4.858916
# 6      pen Pen F 8 8 1.00        6.884332        5.134238
 
MFnest(aCore, 'litter')
#    variable     level N U  MF median_resp:con median_resp:vac
# 1    litter Litter 11 4 4 1.0        8.237194        5.125583
# 2    litter Litter 12 4 2 0.0        4.914294        3.808685
# 3    litter Litter 13 4 4 1.0        8.103441        5.227177
# 4    litter Litter 14 4 4 1.0        8.113648        5.594649
# 5    litter Litter 15 4 4 1.0        8.087914        5.256611
# 6    litter Litter 16 4 4 1.0        6.770457        4.503342
# 7    litter Litter 17 4 4 1.0        5.575649        4.258613
# 8    litter Litter 18 4 3 0.5        7.442926        6.329360
# 9    litter Litter 19 4 4 1.0        7.980120        4.584041
# 10   litter Litter 20 4 4 1.0        6.781480        4.858916
# 11   litter Litter 21 4 4 1.0        6.819483        5.363069
# 12   litter Litter 22 4 4 1.0        7.272879        5.134238

MFnest(aCore, c('room', 'pen', 'litter'))
#    variable     level  N  U        MF median_resp:con median_resp:vac
# 1      room    Room W 24 22 0.8333333        7.792846        4.846574
# 2      room    Room Z 24 23 0.9166667        6.708765        4.978941
# 3       pen     Pen A  8  6 0.5000000        6.792419        4.237541
# 4       pen     Pen B  8  8 1.0000000        8.113648        5.594649
# 5       pen     Pen C  8  8 1.0000000        7.688790        4.846574
# 6       pen     Pen D  8  7 0.7500000        6.097081        4.978941
# 7       pen     Pen E  8  8 1.0000000        6.858526        4.858916
# 8       pen     Pen F  8  8 1.0000000        6.884332        5.134238
# 9    litter Litter 11  4  4 1.0000000        8.237194        5.125583
# 10   litter Litter 12  4  2 0.0000000        4.914294        3.808685
# 11   litter Litter 13  4  4 1.0000000        8.103441        5.227177
# 12   litter Litter 14  4  4 1.0000000        8.113648        5.594649
# 13   litter Litter 15  4  4 1.0000000        8.087914        5.256611
# 14   litter Litter 16  4  4 1.0000000        6.770457        4.503342
# 15   litter Litter 17  4  4 1.0000000        5.575649        4.258613
# 16   litter Litter 18  4  3 0.5000000        7.442926        6.329360
# 17   litter Litter 19  4  4 1.0000000        7.980120        4.584041
# 18   litter Litter 20  4  4 1.0000000        6.781480        4.858916
# 19   litter Litter 21  4  4 1.0000000        6.819483        5.363069
# 20   litter Litter 22  4  4 1.0000000        7.272879        5.134238
}
\seealso{
\code{\link{MFh}}
}

