% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFClusHier.R
\name{MFnest}
\alias{MFnest}
\title{Summations to calculate the MF for nested data from a rank table.}
\usage{
MFnest(Y, which.factor = "All")
}
\arguments{
\item{Y}{rank table (tibble or data.frame), structured as \code{$coreTbl} output from \code{\link{MFh}} or
output list from MFh.}

\item{which.factor}{one or more variable(s) of interest. This can be any of the core or nest variables
from the data set. If none or \code{NULL} is specified, MF will be calculated for the whole 
tree.}
}
\value{
A tibble with each unique level of a variable as a row. Other values 
include: \cr
\describe{
\item{\code{MF}}{Mitigated fraction for the particular level of the variable 
in this row.}
\item{\code{N1N2}}{Sum of the \code{n1n2} variable in \code{$coreTbl} field of 
                  \code{\link{mfhierdata}} object output by \code{\link{MFh}}
                   for this particular variable-level combination.}
\item{\code{U}}{Sum of u variable in \code{$coreTbl} field of 
                  \code{\link{mfhierdata}} object output by \code{\link{MFh}}
                  for this particular variable-level combination.}
\item{\code{_N}}{Sum of the \code{_n} variable in \code{$coreTbl} field of 
\code{\link{mfhierdata}} object output by \code{\link{MFh}}
                   for this particular variable-level combination.}
\item{\code{median_resp:}}{Median of responses for each comparison group for 
this particular variable-level combination.}
}
}
\note{
Core variable is the variable corresponding to the lowest nodes of the hierarchial 
tree. Nest variables are those above the core. All refers to a summary of the entire tree.
}
\examples{
a <- data.frame(
 room = paste('Room',rep(c('W','Z'),each=24)),
 pen = paste('Pen',rep(LETTERS[1:6],each=8)),
 litter = paste('Litter',rep(11:22,each=4)),
 tx = rep(rep(c('vac','con'),each=2),12),
 stringsAsFactors = FALSE
 )
set.seed(76153)
a$lung[a$tx=='vac'] <- rnorm(24,5,1.3)
a$lung[a$tx=='con'] <- rnorm(24,7,1.3)

aCore <- MFh(lung ~ tx + room/pen/litter,a)
MFnest(aCore)
#   # A tibble: 1 x 9
#     variable level    MF  N1N2     U con_N vac_N `median_resp:con`
#     <fct>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl>             <dbl>
#   1 All      All   0.875    48    45    24    24              7.24
# ... with 1 more variable: `median_resp:vac` <dbl>

MFnest(aCore$coreTbl)
# Skipping median summary, no response data provided.
# #   A tibble: 1 x 7
#     variable level    MF  N1N2     U con_N vac_N
#     <fct>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#   1 All      All   0.875    48    45    24    24

MFnest(aCore, 'room')
# # A tibble: 2 x 9
#   variable level     MF  N1N2     U con_N vac_N `median_resp:con`
#   <fct>    <chr>  <dbl> <dbl> <dbl> <dbl> <dbl>             <dbl>
# 1 room     Room W 0.833    24    22    12    12              7.79
# 2 room     Room Z 0.917    24    23    12    12              6.71
# # ... with 1 more variable: `median_resp:vac` <dbl>

MFnest(aCore, 'pen')
# Complete separation observed for variable(s): pen
# # A tibble: 6 x 9
#   variable level    MF  N1N2     U con_N vac_N `median_resp:con`
#   <fct>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl>             <dbl>
# 1 pen      Pen A  0.5      8     6     4     4              6.79
# 2 pen      Pen B  1        8     8     4     4              8.11
# 3 pen      Pen C  1        8     8     4     4              7.69
# 4 pen      Pen D  0.75     8     7     4     4              6.10
# 5 pen      Pen E  1        8     8     4     4              6.86
# 6 pen      Pen F  1        8     8     4     4              6.88
# # ... with 1 more variable: `median_resp:vac` <dbl>

MFnest(aCore, c('All', 'litter'))
# Complete separation observed for variable(s): litter
# # A tibble: 13 x 9
#    variable level        MF  N1N2     U con_N vac_N `median_resp:con`
#    <fct>    <chr>     <dbl> <dbl> <dbl> <dbl> <dbl>             <dbl>
#  1 All      All       0.875    48    45    24    24              7.24
#  2 litter   Litter 11 1         4     4     2     2              8.24
#  3 litter   Litter 12 0         4     2     2     2              4.91
#  4 litter   Litter 13 1         4     4     2     2              8.10
#  5 litter   Litter 14 1         4     4     2     2              8.11
#  6 litter   Litter 15 1         4     4     2     2              8.09
#  7 litter   Litter 16 1         4     4     2     2              6.77
#  8 litter   Litter 17 1         4     4     2     2              5.58
#  9 litter   Litter 18 0.5       4     3     2     2              7.44
# 10 litter   Litter 19 1         4     4     2     2              7.98
# 11 litter   Litter 20 1         4     4     2     2              6.78
# 12 litter   Litter 21 1         4     4     2     2              6.82
# 13 litter   Litter 22 1         4     4     2     2              7.27
# # ... with 1 more variable: `median_resp:vac` <dbl>

MFnest(aCore, 'litter')
# Complete separation observed for variable(s): litter
# #  A tibble: 12 x 9
#    variable level        MF  N1N2     U con_N vac_N `median_resp:con`
#    <fct>    <chr>     <dbl> <dbl> <dbl> <dbl> <dbl>             <dbl>
#  1 litter   Litter 11   1       4     4     2     2              8.24
#  2 litter   Litter 12   0       4     2     2     2              4.91
#  3 litter   Litter 13   1       4     4     2     2              8.10
#  4 litter   Litter 14   1       4     4     2     2              8.11
#  5 litter   Litter 15   1       4     4     2     2              8.09
#  6 litter   Litter 16   1       4     4     2     2              6.77
#  7 litter   Litter 17   1       4     4     2     2              5.58
#  8 litter   Litter 18   0.5     4     3     2     2              7.44
#  9 litter   Litter 19   1       4     4     2     2              7.98
# 10 litter   Litter 20   1       4     4     2     2              6.78
# 11 litter   Litter 21   1       4     4     2     2              6.82
# 12 litter   Litter 22   1       4     4     2     2              7.27
# # ... with 1 more variable: `median_resp:vac` <dbl>

MFnest(aCore, c('room', 'pen', 'litter'))
# Complete separation observed for variable(s): litter, pen
# #  A tibble: 20 x 9
#    variable level        MF  N1N2     U con_N vac_N `median_resp:con`
#    <fct>    <chr>     <dbl> <dbl> <dbl> <dbl> <dbl>             <dbl>
#  1 room     Room W    0.833    24    22    12    12              7.79
#  2 room     Room Z    0.917    24    23    12    12              6.71
#  3 pen      Pen A     0.5       8     6     4     4              6.79
#  4 pen      Pen B     1         8     8     4     4              8.11
#  5 pen      Pen C     1         8     8     4     4              7.69
#  6 pen      Pen D     0.75      8     7     4     4              6.10
#  7 pen      Pen E     1         8     8     4     4              6.86
#  8 pen      Pen F     1         8     8     4     4              6.88
#  9 litter   Litter 11 1         4     4     2     2              8.24
# 10 litter   Litter 12 0         4     2     2     2              4.91
# 11 litter   Litter 13 1         4     4     2     2              8.10
# 12 litter   Litter 14 1         4     4     2     2              8.11
# 13 litter   Litter 15 1         4     4     2     2              8.09
# 14 litter   Litter 16 1         4     4     2     2              6.77
# 15 litter   Litter 17 1         4     4     2     2              5.58
# 16 litter   Litter 18 0.5       4     3     2     2              7.44
# 17 litter   Litter 19 1         4     4     2     2              7.98
# 18 litter   Litter 20 1         4     4     2     2              6.78
# 19 litter   Litter 21 1         4     4     2     2              6.82
# 20 litter   Litter 22 1         4     4     2     2              7.27
# # ... with 1 more variable: `median_resp:vac` <dbl>
}
\seealso{
\code{\link{MFh}}
}
