% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFClusHier.R
\name{MFnest}
\alias{MFnest}
\title{Summations to calculate the MF for nested data from a rank table.}
\usage{
MFnest(Y, which.factor = NULL)
}
\arguments{
\item{Y}{rank table, structured as \code{$coreTbl} output from \code{\link{MFh}} or
output list from MFh}

\item{which.factor}{one or more variable(s) of interest. This can be any of the core or nest variables
from the data set. If none or \code{NULL} is specified, MF will be calculated for the whole 
tree.}
}
\value{
data.frame with each unique level of a variable as a row. Values \code{N}, 
}
\note{
Core variable is the variable corresponding to the lowest nodes of the hierarchial 
tree. Nest variables are those above the core. All refers to a summary of the entire tree.
}
\examples{
MFnest(aCore)
#   variable level  N  U    MF
# 1      All   All 48 45 0.875

MFnest(aCore$coreTbl)
#   variable level  N  U    MF
# 1      All   All 48 45 0.875

MFnest(aCore, 'room')
#  variable  level  N  U        MF
# 1     room Room W 24 22 0.8333333
# 2     room Room Z 24 23 0.9166667

MFnest(aCore, 'pen')
#   variable level N U   MF
# 1      pen Pen A 8 6 0.50
# 2      pen Pen B 8 8 1.00
# 3      pen Pen C 8 8 1.00
# 4      pen Pen D 8 7 0.75
# 5      pen Pen E 8 8 1.00
# 6      pen Pen F 8 8 1.00
 
MFnest(aCore, 'litter')
#    variable     level N U  MF
# 1    litter Litter 11 4 4 1.0
# 2    litter Litter 12 4 2 0.0
# 3    litter Litter 13 4 4 1.0
# 4    litter Litter 14 4 4 1.0
# 5    litter Litter 15 4 4 1.0
# 6    litter Litter 16 4 4 1.0
# 7    litter Litter 17 4 4 1.0
# 8    litter Litter 18 4 3 0.5
# 9    litter Litter 19 4 4 1.0
# 10   litter Litter 20 4 4 1.0
# 11   litter Litter 21 4 4 1.0
# 12   litter Litter 22 4 4 1.0

MFnest(aCore, c('room', 'pen', 'litter'))
#    variable     level  N  U        MF
# 1      room    Room W 24 22 0.8333333
# 2      room    Room Z 24 23 0.9166667
# 3       pen     Pen A  8  6 0.5000000
# 4       pen     Pen B  8  8 1.0000000
# 5       pen     Pen C  8  8 1.0000000
# 6       pen     Pen D  8  7 0.7500000
# 7       pen     Pen E  8  8 1.0000000
# 8       pen     Pen F  8  8 1.0000000
# 9    litter Litter 11  4  4 1.0000000
# 10   litter Litter 12  4  2 0.0000000
# 11   litter Litter 13  4  4 1.0000000
# 12   litter Litter 14  4  4 1.0000000
# 13   litter Litter 15  4  4 1.0000000
# 14   litter Litter 16  4  4 1.0000000
# 15   litter Litter 17  4  4 1.0000000
# 16   litter Litter 18  4  3 0.5000000
# 17   litter Litter 19  4  4 1.0000000
# 18   litter Litter 20  4  4 1.0000000
# 19   litter Litter 21  4  4 1.0000000
# 20   litter Litter 22  4  4 1.0000000
}
\seealso{
\code{\link{MFh}}
}

