% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFClusHier.R
\name{MFnest}
\alias{MFnest}
\title{MFnest}
\usage{
MFnest(Y, which.factor = NULL)
}
\examples{

MFnest(aCore)
MFnest(aCore, 'room')
MFnest(aCore, 'pen')
MFnest(aCore, 'litter')
MFnest(aCore)
#   level  N  U    MF
# 1   All 48 45 0.875

MFnest(aCore, 'room')
#    level  N  U        MF
# 1 Room W 24 22 0.8333333
# 2 Room Z 24 23 0.9166667

MFnest(aCore, 'pen')
#   level N U   MF
# 1 Pen A 8 6 0.50
# 2 Pen B 8 8 1.00
# 3 Pen C 8 8 1.00
# 4 Pen D 8 7 0.75
# 5 Pen E 8 8 1.00
# 6 Pen F 8 8 1.00

MFnest(aCore, 'litter')
#        level N U  MF
# 1  Litter 11 4 4 1.0
# 2  Litter 12 4 2 0.0
# 3  Litter 13 4 4 1.0
# 4  Litter 14 4 4 1.0
# 5  Litter 15 4 4 1.0
# 6  Litter 16 4 4 1.0
# 7  Litter 17 4 4 1.0
# 8  Litter 18 4 3 0.5
# 9  Litter 19 4 4 1.0
# 10 Litter 20 4 4 1.0
# 11 Litter 21 4 4 1.0
# 12 Litter 22 4 4 1.0
}

