% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFClusHier.R
\name{MFnest}
\alias{MFnest}
\title{Summations to calculate the MF for nested data from a rank table.}
\usage{
MFnest(Y, which.factor = "All")
}
\arguments{
\item{Y}{rank table (tibble or data.frame), structured as \code{$coreTbl} output from \code{\link{MFh}} or
output list from MFh.}

\item{which.factor}{one or more variable(s) of interest. This can be any of the core or nest variables
from the data set. If none or \code{NULL} is specified, MF will be calculated for the whole 
tree.}
}
\value{
tibble with each unique level of a variable as a row. Values \code{N}, 
\code{U}, \code{MF} and median of responses returned.
}
\note{
Core variable is the variable corresponding to the lowest nodes of the hierarchial 
tree. Nest variables are those above the core. All refers to a summary of the entire tree.
}
\examples{
MFnest(aCore)
#   # A tibble: 1 x 7
#     variable level     N     U    MF `median_resp:con` `median_resp:vac`
#     <fct>    <chr> <dbl> <dbl> <dbl>             <dbl>             <dbl>
#   1 All      All      48    45 0.875              7.24              4.91

MFnest(aCore$coreTbl)
# Skipping median summary, no response data provided.
# # A tibble: 1 x 5
#   variable level     N     U    MF
#   <fct>    <chr> <dbl> <dbl> <dbl>
# 1 All      All      48    45 0.875

MFnest(aCore, 'room')
# # A tibble: 2 x 7
#   variable level      N     U    MF `median_resp:con` `median_resp:vac`
#   <fct>    <chr>  <dbl> <dbl> <dbl>             <dbl>             <dbl>
# 1 room     Room W    24    22 0.833              7.79              4.85
# 2 room     Room Z    24    23 0.917              6.71              4.98

MFnest(aCore, 'pen')
# Complete separation observed for variable(s): pen
# # A tibble: 6 x 7
#   variable level     N     U    MF `median_resp:con` `median_resp:vac`
#   <fct>    <chr> <dbl> <dbl> <dbl>             <dbl>             <dbl>
# 1 pen      Pen A     8     6  0.5               6.79              4.24
# 2 pen      Pen B     8     8  1                 8.11              5.59
# 3 pen      Pen C     8     8  1                 7.69              4.85
# 4 pen      Pen D     8     7  0.75              6.10              4.98
# 5 pen      Pen E     8     8  1                 6.86              4.86
# 6 pen      Pen F     8     8  1                 6.88              5.13

MFnest(aCore, c('All', 'litter'))
# Complete separation observed for variable(s): litter
# # A tibble: 13 x 7
#  variable level         N     U    MF `median_resp:con` `median_resp:vac`
#  <fct>    <chr>     <dbl> <dbl> <dbl>             <dbl>             <dbl>
#  1 All      All          48    45 0.875              7.24              4.91
#  2 litter   Litter 11     4     4 1                  8.24              5.13
#  3 litter   Litter 12     4     2 0                  4.91              3.81
#  4 litter   Litter 13     4     4 1                  8.10              5.23
#  5 litter   Litter 14     4     4 1                  8.11              5.59
#  6 litter   Litter 15     4     4 1                  8.09              5.26
#  7 litter   Litter 16     4     4 1                  6.77              4.50
#  8 litter   Litter 17     4     4 1                  5.58              4.26
#  9 litter   Litter 18     4     3 0.5                7.44              6.33
# 10 litter   Litter 19     4     4 1                  7.98              4.58
# 11 litter   Litter 20     4     4 1                  6.78              4.86
# 12 litter   Litter 21     4     4 1                  6.82              5.36
# 13 litter   Litter 22     4     4 1                  7.27              5.13

MFnest(aCore, 'litter')
# Complete separation observed for variable(s): litter
# # A tibble: 12 x 7
#    variable level         N     U    MF `median_resp:con` `median_resp:vac`
#    <fct>    <chr>     <dbl> <dbl> <dbl>             <dbl>             <dbl>
#  1 litter   Litter 11     4     4   1                8.24              5.13
#  2 litter   Litter 12     4     2   0                4.91              3.81
#  3 litter   Litter 13     4     4   1                8.10              5.23
#  4 litter   Litter 14     4     4   1                8.11              5.59
#  5 litter   Litter 15     4     4   1                8.09              5.26
#  6 litter   Litter 16     4     4   1                6.77              4.50
#  7 litter   Litter 17     4     4   1                5.58              4.26
#  8 litter   Litter 18     4     3   0.5              7.44              6.33
#  9 litter   Litter 19     4     4   1                7.98              4.58
# 10 litter   Litter 20     4     4   1                6.78              4.86
# 11 litter   Litter 21     4     4   1                6.82              5.36
# 12 litter   Litter 22     4     4   1                7.27              5.13

MFnest(aCore, c('room', 'pen', 'litter'))
# Complete separation observed for variable(s): litter, pen
# # A tibble: 20 x 7
#    variable level         N     U    MF `median_resp:con` `median_resp:vac`
#    <fct>    <chr>     <dbl> <dbl> <dbl>             <dbl>             <dbl>
#  1 room     Room W       24    22 0.833              7.79              4.85
#  2 room     Room Z       24    23 0.917              6.71              4.98
#  3 pen      Pen A         8     6 0.5                6.79              4.24
#  4 pen      Pen B         8     8 1                  8.11              5.59
#  5 pen      Pen C         8     8 1                  7.69              4.85
#  6 pen      Pen D         8     7 0.75               6.10              4.98
#  7 pen      Pen E         8     8 1                  6.86              4.86
#  8 pen      Pen F         8     8 1                  6.88              5.13
#  9 litter   Litter 11     4     4 1                  8.24              5.13
# 10 litter   Litter 12     4     2 0                  4.91              3.81
# 11 litter   Litter 13     4     4 1                  8.10              5.23
# 12 litter   Litter 14     4     4 1                  8.11              5.59
# 13 litter   Litter 15     4     4 1                  8.09              5.26
# 14 litter   Litter 16     4     4 1                  6.77              4.50
# 15 litter   Litter 17     4     4 1                  5.58              4.26
# 16 litter   Litter 18     4     3 0.5                7.44              6.33
# 17 litter   Litter 19     4     4 1                  7.98              4.58
# 18 litter   Litter 20     4     4 1                  6.78              4.86
# 19 litter   Litter 21     4     4 1                  6.82              5.36
# 20 litter   Litter 22     4     4 1                  7.27              5.13
}
\seealso{
\code{\link{MFh}}
}
