% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFnestBoot.r
\name{MFnestBoot}
\alias{MFnestBoot}
\title{MFnestBoot}
\usage{
MFnestBoot(x, which.factor = "All", alpha = 0.05)
}
\arguments{
\item{x}{output from \code{\link{MFhBoot}}}

\item{which.factor}{Which variables to include in the mitigated fraction summation.
Default is ’All’, to sum over entire tree.}

\item{alpha}{Passed to \code{\link[MF]{emp.hpd}} to calculate high tailed upper 
and high tailed lower 
of mitigated fraction}
}
\description{
MFnest using bootstrapping
}
\examples{
set.seed(76153)
a <- data_frame(room = paste('Room', rep(c('W','Z'), each = 24)),
                pen = paste('Pen', rep(LETTERS[1:6], each = 8)),
                litter = paste('Litter', rep(11:22, each = 4)),
                tx = rep(rep(c('vac', 'con'), each = 2), 12)) \%>\%
  mutate(lung = ifelse(tx == 'vac', rnorm(24, 5, 1.3), rnorm(24, 7, 1.3)))
a

formula <- lung ~ tx + room/pen/litter
nboot <- 10000
boot.cluster <- TRUE
boot.unit <- TRUE
which.factors <- c('All', 'room', 'pen', 'litter')

#################
set.seed(12345)

test1 <- MFhBoot(formula, a, 
                 nboot = 10000,
                 boot.cluster = TRUE, boot.unit = TRUE)
MFnestBoot(test1, c('All', 'litter'))

\dontrun{
system.time(test2 <- MFnestBoot(test1, which.factors))
test2
system.time(test3 <- MFnestBoot(test1, which.factors[1]))
test3
system.time(test4 <- MFnestBoot(test1, which.factors[2]))
test4
system.time(test5 <- MFnestBoot(test1, which.factors[2:3]))
test5
system.time(test6 <- MFnestBoot(test1, which.factors[2:4]))
test6
}
}
\seealso{
\code{\link{MFClusBootHier}}, \code{\link{MFhBoot}}
}
