% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF-package.r
\docType{data}
\name{mlesions}
\alias{mlesions}
\alias{mlesions-data}
\title{mlesions dataset}
\format{
A data frame with 52 observations of the following 3 variables, no
NAs.
\itemize{
\item \code{cage}: Cage ID. 1 - 26.
\item \code{tx}: Treatment. One of "con" or "vac".
\item \code{les}: Percent gross lung lesions.
}
}
\description{
Post-mortem examination of the lungs of dogs housed in cages by
pairs.
}
\keyword{datasets}
