\docType{data}
\name{mlesions}
\alias{mlesions}
\title{mlesions dataset}
\format{a data frame with 52 observations of the following 3 variables, no NAs
\describe{
\item{cage}{Cage ID. 1 - 26}
\item{tx}{Treatment. One of 'con' or 'vac'}
\item{les}{Percent lung lesion}
}}
\description{
something here
}
\keyword{datasets}

