% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF-package.r
\docType{data}
\name{piglung}
\alias{piglung}
\alias{piglung-data}
\title{piglung dataset}
\format{
A data frame with 102 observations of the following 3 variables, no
  NAs. \describe{

  \item{lesion}{Percent gross lung lesions.}

  \item{group}{Treatment group. One of 'con' or 'vac'.}

  \item{litter}{Litter ID.}

  }
}
\description{
Post-mortem examination of the lungs of pigs in litters.
}
\keyword{datasets}
